Fixed: size of _SCImageAndStyledTextView was not calculated correctly.
Previously, _CPImageAndTextView was allowed to calculate the size of the view based on the text. Unfortunately it was calculating based on the size of the markup, not on the size of the formatted text.

This commit precalculates the formatted text size (_textSize), which forces _CPImageAndTextView to use that size when positioning the view.